pyproject.toml
[project]
name = "zephyops-blog-content-automation-pipline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.56.0",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
